Roll some time display functions into the rio-tools (rt) package.
