+++ rt/helpers.go
@@ -33,3 +33,19 @@ func Plural[T Ordered](n T, singular, plural string) string {
 	}
 	return plural
 }
+
+// WrapItem wraps a value of any type in a func.
+// This is used to inject values into the template.FuncMap.
+func WrapItem[T any](val T) func() T {
+	return func() T {
+		return val
+	}
+}
+
+// WrapItems wraps a slice of any type in a func.
+// This is used to inject values into the template.FuncMap.
+func WrapItems[T any](vals []T) func() []T {
+	return func() []T {
+		return vals
+	}
+}
